+++ NEWS.md
@@ -1,4 +1,6 @@
 # denim 1.2.1
+* Minor runtime improvement
+* Add option to select color palette for plot
 
 # denim 1.2.0
 * Add denim DSL support